Move controller manager to plugin/
