+++ plugin/cmd/controller-manager/controller-manager.go
@@ -26,9 +26,9 @@ import (
 	"time"
 
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/client"
-	"github.com/GoogleCloudPlatform/kubernetes/pkg/controller"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/util"
 	verflag "github.com/GoogleCloudPlatform/kubernetes/pkg/version/flag"
+	"github.com/GoogleCloudPlatform/kubernetes/plugin/pkg/controller-manager"
 	"github.com/coreos/go-etcd/etcd"
 	"github.com/golang/glog"
 )
