+++ plugin/pkg/controller-manager/doc.go
